debug-example-next/backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.24.0
openai>=1.0.0
aiofiles>=23.0.0
python-multipart>=0.0.6 